container-app/media/requirements.txt
Flask==2.3.3
boto3==1.37.18
python-dotenv==1.1.0
pymongo==4.12.1